<refactor>(all) Optimize import, Line formatting
