+++ src/main/java/kitchenpos/menus/application/TobeMenuService.java
@@ -21,6 +21,7 @@ public class TobeMenuService {
     private final TobeMenuGroupRepository menuGroupRepository;
     private final TobeProductRepository productRepository;
 
+
     public TobeMenuService(
             final TobeMenuRepository menuRepository,
             final TobeMenuGroupRepository menuGroupRepository,
