+++ src/main/java/kitchenpos/menus/domain/tobe/domain/policy/DefaultMenuGroupNamingRule.java
@@ -1,7 +1,7 @@
 package kitchenpos.menus.domain.tobe.domain.policy;
 
-import kitchenpos.common.exception.NamingRuleViolationException;
-import kitchenpos.common.policy.NamingRule;
+import kitchenpos.support.exception.NamingRuleViolationException;
+import kitchenpos.support.policy.NamingRule;
 
 import java.util.Objects;
