+++ src/main/java/kitchenpos/menus/domain/tobe/domain/policy/DefaultMenuNamingRule.java
@@ -1,7 +1,7 @@
 package kitchenpos.menus.domain.tobe.domain.policy;
 
-import kitchenpos.common.exception.NamingRuleViolationException;
-import kitchenpos.common.policy.NamingRule;
+import kitchenpos.support.exception.NamingRuleViolationException;
+import kitchenpos.support.policy.NamingRule;
 import kitchenpos.products.infra.PurgomalumClient;
 
 import java.util.Objects;
