+++ src/main/java/kitchenpos/support/exception/PricingRuleViolationException.java
@@ -1,4 +1,4 @@
-package kitchenpos.common.exception;
+package kitchenpos.support.exception;
 
 public class PricingRuleViolationException extends RuntimeException {
