+++ src/main/java/kitchenpos/support/policy/NamingRule.java
@@ -1,4 +1,4 @@
-package kitchenpos.common.policy;
+package kitchenpos.support.policy;
 
 public interface NamingRule {
     boolean checkRule(String name);
