+++ src/test/java/kitchenpos/products/application/TobeProductServiceTest.java
@@ -1,11 +1,11 @@
 package kitchenpos.products.application;
 
-import kitchenpos.common.exception.NamingRuleViolationException;
-import kitchenpos.common.exception.PricingRuleViolationException;
-import kitchenpos.common.policy.FakeFailNamingRule;
-import kitchenpos.common.policy.FakeFailPricingRule;
-import kitchenpos.common.policy.FakeSuccessNamingRule;
-import kitchenpos.common.policy.FakeSuccessPricingRule;
+import kitchenpos.support.exception.NamingRuleViolationException;
+import kitchenpos.support.exception.PricingRuleViolationException;
+import kitchenpos.support.policy.FakeFailNamingRule;
+import kitchenpos.support.policy.FakeFailPricingRule;
+import kitchenpos.support.policy.FakeSuccessNamingRule;
+import kitchenpos.support.policy.FakeSuccessPricingRule;
 import kitchenpos.menus.application.InMemoryMenuRepository;
 import kitchenpos.menus.domain.MenuRepository;
 import kitchenpos.products.domain.tobe.domain.InMemoryTobeProductRepository;
