+++ src/test/java/kitchenpos/support/policy/FakeSuccessNamingRule.java
@@ -1,4 +1,4 @@
-package kitchenpos.common.policy;
+package kitchenpos.support.policy;
 
 public class FakeSuccessNamingRule implements NamingRule {
     @Override
